daemon: do not wait for result from RegenerateWait when endpoint is added to endpointmanager
If the entire regeneration for an endpoint is waited to be completed (e.g.,
policy generation, BPF compilation / loading, and proxy updates) when an
endpoint is created, then creation of endpoints which have L7 policy which
selects them when Cilium is running with Istio will result in a deadlock. This
is because when a pod is created, its sidecar proxy is not started yet, so when
addition of the endpoint into the endpoint manager waits for regeneration to
complete, the regeneration gets stuck because the its trying to configure a
proxy which does not exist it. To get around this, do not synchronously wait for
the entire regeneration process to succeed for a given endpoint in
endpointmanager. Instead, launch a goroutine which waits for the result of the
first regeneration, and if that regeneration fails, then remove the endpoint
from the endpointmanager. When the Cilium CNI plugin creates an endpoint, the
API call will still wait synchronously for the generation of policy and loading
of the endpoint's BPF program, but not the entire regeneration process which
includes proxy updates, to complete. This means that initial L3/L4 connectivity
is only in place when endpoint creation is completed when CNI creates an
endpoint.

Signed-off by: Ian Vernon <ian@cilium.io>